Challenge #3
Siderophores database:

This database is made up of 221 siderophores and includes the name of the compound, the associated microorganism, the type and number of chelating groups it contains (such as hydroxamate, catecholate, a-hydroxycarboxylate, carboxylate, phenolate, citrate, other* ), and the molecular representations SMILES, SMARTS and SELFIES (Siderophore_DB.csv). Descriptors: The files Siderophore_DB1, Siderophore_DB2 and Siderophore_DB3 include molecular RDKit (119), Morgan fingerprints (2049) and Mordred (1827), respectively.